consumer/requirements.txt
kafka-python
psycopg2-binary
pandas
confluent-kafka
pandas
sqlalchemy 
